requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dlib @ file:///C:/Users/Aksa/OneDrive/Desktop/Python_project/dlib
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.1
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
six==1.17.0
Werkzeug==3.1.3
